world-manager is responsible of storing the world runtime metadata when new scene is deployed
